Module rascaline::calculators

Re-exports

pub use self::dummy_calculator::DummyCalculator;

pub use self::soap::SphericalExpansionByPair;

pub use self::soap::SphericalExpansionParameters;

pub use self::soap::SphericalExpansion;

pub use self::soap::SoapPowerSpectrum;

pub use self::soap::PowerSpectrumParameters;

pub use self::soap::SoapRadialSpectrum;

pub use self::soap::RadialSpectrumParameters;

pub use self::lode::LodeSphericalExpansion;

pub use self::lode::LodeSphericalExpansionParameters;

Modules

lode

soap

Structs

AtomicComposition

An atomic composition calculator for obtaining the stoichiometric information.

GtoRadialBasis

Use a radial basis similar to Gaussian-Type Orbitals.

NeighborList

This calculator computes the neighbor list for a given spherical cutoff, and returns the list of distance vectors between all pairs of atoms strictly inside the cutoff.

SortedDistances

Sorted distances vector representation of an atomic environment.

Enums

RadialBasis

Radial basis that can be used in the SOAP or LODE spherical expansion

Traits

CalculatorBase

The CalculatorBase trait is the interface shared by all calculator implementations; and used by crate::Calculator to run the calculation.