Module rascaline::calculators::lode

Structs

LodeRadialIntegralGto

Implementation of the LODE radial integral for GTO radial basis and Gaussian atomic density.

LodeRadialIntegralGtoParameters

Parameters controlling the LODE radial integral with GTO radial basis

LodeRadialIntegralParameters

Parameters controlling the radial integral for LODE

LodeRadialIntegralSpline

LodeRadialIntegralSpline allows to evaluate another radial integral implementation using cubic Hermit spline.

LodeRadialIntegralSplineParameters

Parameters for computing the radial integral using Hermit cubic splines

LodeSphericalExpansion

The actual calculator used to compute LODE spherical expansion coefficients

LodeSphericalExpansionParameters

Parameters for LODE spherical expansion calculator.

Traits

LodeRadialIntegral

A LodeRadialIntegral computes the LODE radial integral on a given radial basis.