Trait rascaline::calculators::lode::LodeRadialIntegral

pub trait LodeRadialIntegral: RefUnwindSafe + Send {
    // Required methods
    fn compute(
        &self,
        k_norm: f64,
        values: ArrayViewMut2<'_, f64>,
        gradients: Option<ArrayViewMut2<'_, f64>>,
    );
    fn compute_center_contribution(&self) -> Array1<f64>;
}

A LodeRadialIntegral computes the LODE radial integral on a given radial basis.

See equations 5 to 8 of this paper for mor information on the radial integral.

std::panic::RefUnwindSafe is a required super-trait to enable passing radial integrals across the C API. Send is a required super-trait to enable passing radial integrals between threads.

Required Methods

fn compute( &self, k_norm: f64, values: ArrayViewMut2<'_, f64>, gradients: Option<ArrayViewMut2<'_, f64>>, )

Compute the LODE radial integral for a single k-vector norm and store the resulting data in the (max_angular + 1) x max_radial array values. If gradients is Some, also compute and store gradients there.

fn compute_center_contribution(&self) -> Array1<f64>

Compute the contribution of the central atom to the final <n l m> coefficients. By symmetry, only l=0 is non-zero, so this function returns a 1-D array containing the different <n 0 0> coefficients.

This function differs from the rest of LODE calculation because it goes straight from atom => n l m, without using k-space projection in the middle.