Struct rascaline::calculators::soap::SphericalExpansionByPair

pub struct SphericalExpansionByPair { /* private fields */ }

The actual calculator used to compute spherical expansion pair-by-pair

Implementations

impl SphericalExpansionByPair

pub fn new( parameters: SphericalExpansionParameters ) -> Result<SphericalExpansionByPair, Error>

pub fn parameters(&self) -> &SphericalExpansionParameters

Access the spherical expansion parameters used by this calculator