Trait rascaline::calculators::soap::RadialIntegral

pub trait RadialIntegral: RefUnwindSafe + Send {
    fn compute(
        &self, 
        rij: f64, 
        values: ArrayViewMut2<'_, f64>, 
        gradients: Option<ArrayViewMut2<'_, f64>>
    );
}

A RadialIntegral computes the radial integral on a given radial basis.

See equations 5 to 8 of this paper for mor information on the radial integral.

std::panic::RefUnwindSafe is a required super-trait to enable passing radial integrals across the C API. Send is a required super-trait to enable passing radial integrals between threads.

Required methods

fn compute(
&self,
rij: f64,
values: ArrayViewMut2<'_, f64>,
gradients: Option<ArrayViewMut2<'_, f64>>
)

Compute the radial integral for a single distance between two atoms and store the resulting data in the max_radial x (max_angular + 1) array values. If gradients is Some, also compute and store gradients there.