Trait rascaline::calculators::soap::SoapRadialIntegral

pub trait SoapRadialIntegral: RefUnwindSafe + Send {
    // Required method
    fn compute(
        &self,
        rij: f64,
        values: ArrayViewMut2<'_, f64>,
        gradients: Option<ArrayViewMut2<'_, f64>>
    );
}

A SoapRadialIntegral computes the SOAP radial integral on a given radial basis.

See equations 5 to 8 of this paper for mor information on the radial integral.

std::panic::RefUnwindSafe is a required super-trait to enable passing radial integrals across the C API. Send is a required super-trait to enable passing radial integrals between threads.

Required Methods

fn compute( &self, rij: f64, values: ArrayViewMut2<'_, f64>, gradients: Option<ArrayViewMut2<'_, f64>> )

Compute the radial integral for a single distance between two atoms and store the resulting data in the (max_angular + 1) x max_radial array values. If gradients is Some, also compute and store gradients there.

The radial integral $I_{nl}$ is defined as “the non-spherical harmonics part of the spherical expansion”. Depending on the atomic density, different expressions can be used.

For a delta density, the radial integral is simply the radial basis function $R_{nl}$ evaluated at the pair distance:

$$ I_{nl}(r_{ij}) = R_{nl}(r_{ij}) $$

For a Gaussian atomic density with a width of $\sigma$, the radial integral reduces to:

$$ I_{nl}(r_{ij}) = \frac{4\pi}{(\pi \sigma^2)^{3/4}} e^{-\frac{r_{ij}^2}{2\sigma^2}} \int_0^\infty \mathrm{d}r r^2 R_{nl}(r) e^{-\frac{r^2}{2\sigma^2}} i_l\left(\frac{rr_{ij}}{\sigma^2}\right) $$

where $i_l$ is the modified spherical Bessel function of the first kind of order $l$.

Finally, for an arbitrary spherically symmetric atomic density g(r), the radial integral is

$$ I_{nl}(r_{ij}) = 2\pi \int_0^\infty \mathrm{d}r r^2 R_{nl}(r) \int_{-1}^1 \mathrm{d}u P_l(u) g(\sqrt{r^2+r_{ij}^2-2rr_{ij}u}) $$

where $P_l$ is the l-th Legendre polynomial.