Trait rascaline::descriptor::SamplesBuilder

pub trait SamplesBuilder {
    fn names(&self) -> Vec<&str>;

    fn gradients_names(&self) -> Option<Vec<&str>>;

    fn samples(&self, systems: &mut [Box<dyn System>]) -> Result<Indexes, Error>;

    fn with_gradients(
        &self, 
        systems: &mut [Box<dyn System>]
    ) -> Result<(Indexes, Option<Indexes>), Error> { ... }

    fn gradients_for(
        &self, 
        systems: &mut [Box<dyn System>], 
        samples: &Indexes
    ) -> Result<Option<Indexes>, Error> { ... }
}

The SampleBuilder trait is used to abstract over the different kinds of samples used in Rascaline.

Different implementations of this trait correspond to different types of samples (for example one sample for each structure; or one sample for each pair, etc.)

Each implementation must be able to generate samples from a list of systems, and can optionally implement samples for gradients calculation.

Required methods

fn names(&self) -> Vec<&str>

The list of names used by the sample indexes

fn gradients_names(&self) -> Option<Vec<&str>>

The list of names used by the sample gradients indexes, if any

fn samples(&self, systems: &mut [Box<dyn System>]) -> Result<Indexes, Error>

Create an Indexes instance containing all the samples corresponding to the given list of systems.

Loading content...
Provided methods

fn with_gradients(
&self,
systems: &mut [Box<dyn System>]
) -> Result<(Indexes, Option<Indexes>), Error>

Create an Indexes instance containing all the samples corresponding to the given list of systems, and optionally another Indexes instance containing the gradient samples corresponding to all the samples indexes.

This function is implemented by default using SampleBuilder::samples and SampleBuilder::gradients_for, there is no need to override it.

fn gradients_for(
&self,
systems: &mut [Box<dyn System>],
samples: &Indexes
) -> Result<Option<Indexes>, Error>

Create an Indexes instance containing the gradient samples corresponding to the given samples in the given systems; and only these.

If this SampleBuilder does not implement gradient, it should return None (that’s the default implementation).

Loading content...