Trait rascaline::system::System

pub trait System {
    fn cell(&self) -> UnitCell;

    fn size(&self) -> usize;

    fn species(&self) -> &[usize];

    fn positions(&self) -> &[Vector3D];

    fn compute_neighbors(&mut self, cutoff: f64);

    fn pairs(&self) -> &[Pair];

    fn pairs_containing(&self, center: usize) -> &[Pair];
}

A System deals with the storage of atoms and related information, as well as the computation of neighbor lists.

Required methods

fn cell(&self) -> UnitCell

Get the unit cell for this system

fn size(&self) -> usize

Get the number of atoms in this system

fn species(&self) -> &[usize]

Get the atomic species for all atoms in this system. The returned value must be a slice of length self.size(), where each different atomic species is identified with a different usize value. These values are usually the atomic number, but don't have to.

fn positions(&self) -> &[Vector3D]

Get the positions for all atoms in this system. The returned value must be a slice of length self.size() containing the cartesian coordinates of all atoms in the system.

fn compute_neighbors(&mut self, cutoff: f64)

Compute the neighbor list according to the given cutoff, and store it for later access with pairs or pairs_around.

fn pairs(&self) -> &[Pair]

Get the list of pairs in this system. This list of pair should only contain each pair once (and not twice as i-j and j-i), should not contain self pairs (i-i); and should only contains pairs where the distance between atoms is actually bellow the cutoff passed in the last call to compute_neighbors. This function is only valid to call after a call to compute_neighbors.

fn pairs_containing(&self, center: usize) -> &[Pair]

Get the list of pairs in this system which include the atom at index center. The same restrictions on the list of pairs as System::pairs applies, with the additional condition that the pair i-j should be included both in the return of pairs_containing(i) and pairs_containing(j).

Loading content...