Module rascaline::systems

Structs

NeighborsList
A neighbor list implementation usable with any system

Pair
Pair of atoms coming from a neighbor list.

SimpleSystem
A simple implementation of System to use when no other is available

UnitCell
An UnitCell defines the system physical boundaries.

Traits

System
A System deals with the storage of atoms and related information, as well as the computation of neighbor lists.

Functions

read_from_file
Read all structures in the file at the given path using chemfiles, and convert them to SimpleSystems.