Function rascaline::systems::read_from_file

pub fn read_from_file(
    path: impl AsRef<Path>,
) -> Result<Vec<SimpleSystem>, Error>

Read all structures in the file at the given path using chemfiles, and convert them to SimpleSystems.

This function can read all formats supported by chemfiles.