Struct rascaline::systems::SimpleSystem

pub struct SimpleSystem { /* private fields */ }

A simple implementation of System to use when no other is available

Implementations

impl SimpleSystem

pub fn new(cell: UnitCell) -> SimpleSystem

Create a new empty system with the given unit cell

pub fn add_atom(&mut self, atomic_type: i32, position: Vector3D)

Add an atom with the given atomic type and position to this system