Struct rascaline::types::Vector3D

#[repr(transparent)]
pub struct Vector3D(_);

A 3-dimensional vector type

A Vector3D implement all the arithmetic operations:

let u = Vector3D::new(1.0, 2.0, 3.0);
let v = Vector3D::new(4.0, -2.0, 1.0);

// Indexing
assert_eq!(u[0], 1.0);
assert_eq!(u[1], 2.0);
assert_eq!(u[2], 3.0);

// Addition
let w = u + v;
assert_eq!(w, Vector3D::new(5.0, 0.0, 4.0));

// Subtraction
let w = u - v;
assert_eq!(w, Vector3D::new(-3.0, 4.0, 2.0));

// Negation
let w = -u;
assert_eq!(w, Vector3D::new(-1.0, -2.0, -3.0));

// Cross product
let w = u ^ v;
assert_eq!(w, Vector3D::new(8.0, 11.0, -10.0));

// Multiplication
let w = 2.0 * u;
assert_eq!(w, Vector3D::new(2.0, 4.0, 6.0));

let w = u * 3.0;
assert_eq!(w, Vector3D::new(3.0, 6.0, 9.0));

// Division
let w = u / 2.0;
assert_eq!(w, Vector3D::new(0.5, 1.0, 1.5));

// Dot product
let a = u * v;
assert_eq!(a, 3.0);

Implementations

impl Vector3D

pub fn new(x: f64, y: f64, z: f64) -> Vector3D

Create a new Vector3D with components x, y, z

Examples

let vector = Vector3D::new(1.0, 0.0, -42.0);

assert_eq!(vector[0], 1.0);
assert_eq!(vector[1], 0.0);
assert_eq!(vector[2], -42.0);

pub fn zero() -> Vector3D

Create a new Vector3D with components 0, 0, 0

Examples

let vector = Vector3D::zero();
assert_eq!(vector[0], 0.0);
assert_eq!(vector[1], 0.0);
assert_eq!(vector[2], 0.0);

pub fn norm2(&self) -> f64

Return the squared euclidean norm of a Vector3D

Examples

let vec = Vector3D::new(1.0, 0.0, -4.0);
assert_eq!(vec.norm2(), 17.0);

pub fn norm(&self) -> f64

Return the euclidean norm of a Vector3D

Examples

let vec = Vector3D::new(1.0, 0.0, -4.0);
assert_eq!(vec.norm(), f64::sqrt(17.0));

pub fn normalized(&self) -> Vector3D

Normalize a Vector3D.

Examples

let vec = Vector3D::new(1.0, 0.0, -4.0);
let n = vec.normalized();
assert_eq!(n.norm(), 1.0);

pub fn tensorial(&self, other: &Vector3D) -> Matrix3

Tensorial product between vectors. The tensorial product between the vectors a and b creates a Matrix3 with component (i, j) equals to a[i] * b[j].

Examples

let a = Vector3D::new(1.0, 0.0, -4.0);
let b = Vector3D::new(1.0, 2.0, 3.0);
let matrix = Matrix3::new([
    [1.0, 2.0, 3.0],
    [0.0, 0.0, 0.0],
    [-4.0, -8.0, -12.0]
]);
assert_eq!(a.tensorial(&b), matrix);

pub fn min(&self) -> f64

Get the minimal value in this vector, using std::f64::min for comparison.

Examples

let vector = Vector3D::new(1.0, 0.0, -4.0);

assert_eq!(vector.min(), -4.0);

pub fn max(&self) -> f64

Get the maximal value in this vector, using std::f64::max for comparison.

Examples

let vector = Vector3D::new(1.0, 0.0, -4.0);

assert_eq!(vector.max(), 1.0);