Crate rascaline

Re-exports

pub use types::*;

pub use systems::System;

Modules

calculators

descriptor

systems

types

This module provides 3D vectors and matrix to be used in all other modules.

Structs

CalculationOptions

Parameters specific to a single call to compute

Calculator

Descriptor

A Descriptor contains the representation of atomistic systems, as computed by a crate::Calculator.

SimpleSystem

A simple implementation of System to use when no other is available

Enums

Error

SelectedIndexes

List of pre-selected indexes on which the user wants to run a calculation