Module rascaline::calculators::soap

Structs

GTO

GTOParameters
Parameters controlling GTO radial basis

HyperGeometricParameters

HyperGeometricSphericalExpansion
Computes the G function and its derivative for all possible values of l < l_max + 1 and n < n_max with a = (n + l + 3) / 2 and b = l + 3/2 using recurrence relationships between different n/l values to speedup the full calculation.

SphericalExpansion

SphericalExpansionParameters

SphericalHarmonics
Compute a full set of spherical harmonics at given positions

SphericalHarmonicsArray
Array storing data for 0 <= l <= l_max, -l <= m <= l. This type implements Index<[isize; 2]> and IndexMut<[isize; 2]> to allow writing code like

Enums

CutoffFunction
Possible values for the smoothing cutoff function

RadialBasis

Traits

RadialIntegral