Module rascaline::calculators::soap

Structs

PowerSpectrumParameters

Parameters for SOAP power spectrum calculator.

RadialSpectrumParameters

Parameters for the SOAP radial spectrum calculator.

SoapPowerSpectrum

Calculator implementing the Smooth Overlap of Atomic Position (SOAP) power spectrum representation of atomistic systems.

SoapRadialIntegralCache

Store together a Radial integral implementation and cached allocation for values/gradients.

SoapRadialIntegralGto

Implementation of the radial integral for GTO radial basis and gaussian atomic density.

SoapRadialIntegralGtoParameters

Parameters controlling the SOAP radial integral with GTO radial basis

SoapRadialIntegralParameters

Parameters controlling the radial integral for SOAP

SoapRadialIntegralSpline

SoapRadialIntegralSpline allows to evaluate another radial integral implementation using cubic Hermit spline.

SoapRadialIntegralSplineParameters

Parameters for computing the radial integral using Hermit cubic splines

SoapRadialSpectrum

Calculator implementing the Radial spectrum representation of atomistic systems.

SphericalExpansion

The actual calculator used to compute SOAP spherical expansion coefficients

SphericalExpansionByPair

The actual calculator used to compute spherical expansion pair-by-pair

SphericalExpansionParameters

Parameters for spherical expansion calculator.

Enums

CutoffFunction

Possible values for the smoothing cutoff function

RadialScaling

Implemented options for radial scaling of the atomic density around an atom

Traits

SoapRadialIntegral

A SoapRadialIntegral computes the SOAP radial integral on a given radial basis.