Trait rascaline::calculators::soap::RadialIntegral

pub trait RadialIntegral: RefUnwindSafe {
    fn compute(
        &self, 
        rij: f64, 
        values: ArrayViewMut2<'_, f64>, 
        gradients: Option<ArrayViewMut2<'_, f64>>
    );
}

Required methods

fn compute(
&self,
rij: f64,
values: ArrayViewMut2<'_, f64>,
gradients: Option<ArrayViewMut2<'_, f64>>
)

Compute the radial integral for a single atomic distance and store the resulting data in the max_radial x max_angular array values. If gradients is Some, also compute and store gradients.

Loading content...