Trait rascaline::calculators::CalculatorBase

pub trait CalculatorBase: RefUnwindSafe {
    fn name(&self) -> String;

    fn get_parameters(&self) -> String;

    fn features_names(&self) -> Vec<&str>;

    fn features(&self) -> Indexes;

    fn samples_builder(&self) -> Box<dyn SamplesBuilder>;

    fn compute_gradients(&self) -> bool;

    fn check_features(&self, indexes: &Indexes) -> Result<(), Error>;

    fn compute(
        &mut self, 
        systems: &mut [Box<dyn System>], 
        descriptor: &mut Descriptor
    ) -> Result<(), Error>;

    fn check_samples(
        &self, 
        indexes: &Indexes, 
        systems: &mut [Box<dyn System>]
    ) -> Result<(), Error> { ... }
}

TODO: docs

std::panic::RefUnwindSafe is a required super-trait to enable passing calculators across the C API.

Required methods

fn name(&self) -> String

Get the name of this Calculator

fn get_parameters(&self) -> String

Get the parameters used to create this Calculator as a JSON string

fn features_names(&self) -> Vec<&str>

Get the names of features for this Calculator

fn features(&self) -> Indexes

Get the default set of features for this Calculator

fn samples_builder(&self) -> Box<dyn SamplesBuilder>

Get the default sample builder for this Calculator

fn compute_gradients(&self) -> bool

Does this calculator compute gradients?

fn check_features(&self, indexes: &Indexes) -> Result<(), Error>

Check that the given indexes are valid feature indexes for this Calculator. This is used by to ensure only valid features are requested

fn compute(
&mut self,
systems: &mut [Box<dyn System>],
descriptor: &mut Descriptor
) -> Result<(), Error>

Core implementation of the descriptor.

This function should compute the descriptor only for samples in descriptor.samples and computing only features in descriptor.features. By default, these would correspond to the samples and features coming from Descriptor::samples() and Descriptor::features() respectively; but the user can request only a subset of them.

Provided methods

fn check_samples(
&self,
indexes: &Indexes,
systems: &mut [Box<dyn System>]
) -> Result<(), Error>

Check that the given indexes are valid samples indexes for this Calculator. This is used by to ensure only valid samples are requested

The default implementation recompute the full set of samples using Self::samples(), and check that all requested samples are part of the full sample set.