Module rascaline::labels

Structs

AtomCenteredSamples

SampleBuilder for atom-centered representation. This will create one sample for each atom, optionally filtering on the central atom species. The samples names are (“structure”, “center”).

CenterSingleNeighborsSpeciesKeys

Compute a set of keys with two variables: the central atom species and a single neighbor atom species.

CenterSpeciesKeys

Compute a set of keys with a single variable, the central atom species.

CenterTwoNeighborsSpeciesKeys

Compute a set of keys with three variables: the central atom species and two neighbor atom species.

Enums

SpeciesFilter

Atomic species filters to be used when building samples and gradient sample

Traits

KeysBuilder

Common interface to create a set of equistore’s TensorMap keys from systems

SamplesBuilder

Abstraction over the different kinds of samples used in rascaline.