Struct rascaline::systems::UnitCell

pub struct UnitCell { /* private fields */ }

An UnitCell defines the system physical boundaries.

The shape of the cell can be any of the CellShape, and will influence how periodic boundary conditions are applied.

Implementations

impl UnitCell

pub fn infinite() -> UnitCell

Create an infinite unit cell

pub fn orthorhombic(a: f64, b: f64, c: f64) -> UnitCell

Create an orthorhombic unit cell, with side lengths a, b, c.

pub fn cubic(length: f64) -> UnitCell

Create a cubic unit cell, with side lengths length, length, length.

pub fn triclinic(
a: f64,
b: f64,
c: f64,
alpha: f64,
beta: f64,
gamma: f64
) -> UnitCell

Create a triclinic unit cell, with side lengths a, b, c and angles alpha, beta, gamma.

pub fn shape(&self) -> CellShape

Get the cell shape

pub fn is_infinite(&self) -> bool

Check if this unit cell is infinite, i.e. if it does not have periodic boundary conditions.

pub fn a(&self) -> f64

Get the first length of the cell (i.e. the norm of the first vector of the cell)

pub fn b(&self) -> f64

Get the second length of the cell (i.e. the norm of the second vector of the cell)

pub fn c(&self) -> f64

Get the third length of the cell (i.e. the norm of the third vector of the cell)

pub fn distances_between_faces(&self) -> Vector3D

Get the distances between faces of the unit cell

pub fn alpha(&self) -> f64

Get the first angle of the cell

pub fn beta(&self) -> f64

Get the second angle of the cell

pub fn gamma(&self) -> f64

Get the third angle of the cell

pub fn volume(&self) -> f64

Get the volume of the cell

pub fn matrix(&self) -> Matrix3

Get the matricial representation of the unit cell

pub fn fractional(&self, vector: Vector3D) -> Vector3D

Get the fractional representation of the vector in this cell

pub fn cartesian(&self, fractional: Vector3D) -> Vector3D

Get the Cartesian representation of the fractional vector in this cell