Trait rascaline::systems::System

pub trait System {
    fn cell(&self) -> Result<UnitCell, Error>;

    fn size(&self) -> Result<usize, Error>;

    fn species(&self) -> Result<&[i32], Error>;

    fn positions(&self) -> Result<&[Vector3D], Error>;

    fn compute_neighbors(&mut self, cutoff: f64) -> Result<(), Error>;

    fn pairs(&self) -> Result<&[Pair], Error>;

    fn pairs_containing(&self, center: usize) -> Result<&[Pair], Error>;
}

A System deals with the storage of atoms and related information, as well as the computation of neighbor lists.

Required methods

fn cell(&self) -> Result<UnitCell, Error>

Get the unit cell for this system

fn size(&self) -> Result<usize, Error>

Get the number of atoms in this system

fn species(&self) -> Result<&[i32], Error>

Get the atomic species for all atoms in this system. The returned value must be a slice of length self.size(), where each different atomic species is identified with a different integer value. These values are usually the atomic number, but don’t have to.

fn positions(&self) -> Result<&[Vector3D], Error>

Get the positions for all atoms in this system. The returned value must be a slice of length self.size() containing the cartesian coordinates of all atoms in the system.

fn compute_neighbors(&mut self, cutoff: f64) -> Result<(), Error>

Compute the neighbor list according to the given cutoff, and store it for later access with pairs or pairs_around.

fn pairs(&self) -> Result<&[Pair], Error>

Get the list of pairs in this system. This list of pair should only contain each pair once (and not twice as i-j and j-i), should not contain self pairs (i-i); and should only contains pairs where the distance between atoms is actually bellow the cutoff passed in the last call to compute_neighbors. This function is only valid to call after a call to compute_neighbors.

fn pairs_containing(&self, center: usize) -> Result<&[Pair], Error>

Get the list of pairs in this system which include the atom at index center. The same restrictions on the list of pairs as System::pairs applies, with the additional condition that the pair i-j should be included both in the return of pairs_containing(i) and pairs_containing(j).